requirements.txt
pycocotools
requests
aiohttp
fpdf
pillow>=10